+++ CHANGES.rst
@@ -5,6 +5,8 @@ unreleased
 -------
 
 - [cleanup] Drop support for python versions older than 3.6
+- [fix] Adjust for mirakuru 2.2.0 and up
+
 
 2.0.0
 -------